requirements.txt
allure_python_commons==2.13.2
Appium_Python_Client==2.11.1
func_timeout==4.3.5
#jmespath==1.0.1
loguru==0.7.2
openpyxl==3.1.2
paho_mqtt==1.6.1
pytest==7.0.0
PyYAML==6.0.1
Requests==2.27.0
retrying==1.3.4
selenium==3.14.0
zmail==0.2.8
pytest-rerunfailures==10.3